Separate out navigation objects for clarity
