+++ studio/schemas/objects/navigation/footerNavItem.js
