+++ studio/schemas/objects/navigation/theme.js
