+++ studio/schemas/schema.js
@@ -6,18 +6,18 @@ import schemaTypes from 'all:part:@sanity/base/schema-type'
 
 // We import object and document schemas
 import blockContent from './objects/blockContent'
-import focusLinkItem from './objects/focusLinkListItem'
-import focusNav from './objects/focusNav'
-import footerNav from './objects/footerNav'
-import footerNavItem from './objects/footerNavItem'
-import minorLinkListItem from './objects/minorLinkListItem'
-import nav from './objects/nav'
-import primaryNav from './objects/primaryNav'
-import primaryNavItem from './objects/primaryNavItem'
+import focusLinkItem from './objects/navigation/focusLinkListItem'
+import focusNav from './objects/navigation/focusNav'
+import footerNav from './objects/navigation/footerNav'
+import footerNavItem from './objects/navigation/footerNavItem'
+import minorLinkListItem from './objects/navigation/minorLinkListItem'
+import nav from './objects/navigation/nav'
+import primaryNav from './objects/navigation/primaryNav'
+import primaryNavItem from './objects/navigation/primaryNavItem'
 import siteMetadata from './objects/siteMetadata'
 import social from './objects/social'
-import subNav from './objects/subNav'
-import theme from './objects/theme'
+import subNav from './objects/navigation/subNav'
+import theme from './objects/navigation/theme'
 import category from './documents/category'
 import post from './documents/post'
 import author from './documents/author'
